feat: Implement image blurring feature
This commit introduces an image blurring feature to the app.

The `ImageViewFilter` composable is added, which uses a
TensorFlow Lite model to detect inappropriate content in images
and blur it if necessary. This composable is now used in
`CategoryItem` and `HomeCard` to moderate displayed images.

Also, updated build configurations to include necessary
dependencies and configurations for the image blurring feature.