Feature: Hero Carousel Component (#307)
* feat: add material 3 carousel package version 1.4.0-alpha18

* feat: add HeroCarousel composable

This commit introduces a new `HeroCarousel` composable function.

The `HeroCarousel` displays a scrollable list of items with one prominent "hero" item and several smaller items, adapting dynamically to the available space. It also includes a partially visible item at the end.

Helper extension functions `isHero` and `heroRatio` have been added to `CarouselItemScope` to determine if an item is the hero item and its relative size.

* refactor: hoist carouselState to be accepted as a parameter in HeroCarousel