Hotfix/popular mapping (#489)
* feat: add PopularUiMedia and its mapping functions

* Refactor: Combine popular movie and TV show initialization

- Combined `initializePopularMovies` and `initializePopularTvShows` into `initializePopularMedia`.

* fix: update popular media handling and improve UI responsiveness

* refactor: migrate PopularSection to use PopularUiMedia

* ci: Use workspace path for Kover report in CI

This commit updates the Kover report path in the CI workflow to use `github.workspace`. This ensures that the report path is correctly resolved, regardless of the runner's environment.

* Fix(ci): Configure Kover variant and adjust report paths

This commit configures a "custom" Kover variant that includes "jvm" and "debug" build types. It also updates the paths for downloading and accessing the Kover coverage report in the GitHub Actions workflow.

---------

Co-authored-by: london-bot[bot] <220529784+london-bot[bot]@users.noreply.github.com>
Co-authored-by: Ahmed Mones <68463985+Al-Taie@users.noreply.github.com>