feat(list): add edit list bottom sheet
This commit introduces the `EditListBottomSheet` composable, which allows users to edit the name of a list.

The bottom sheet includes:
- A title "Edit list".
- A text field for the list name with a leading icon.
- A "Save" button to apply changes.
- A close button.

String resources for "Edit list", "Save", and "List title" have been added in both English and Arabic.
A new drawable icon `editlisticon.xml` has been added for the list name text field.

The following files related to the previous `ViewListItems` feature have been removed as this functionality is being replaced or refactored:
- `ItemsType.kt`
- `MediaUi.kt`
- `ViewListItemsContract.kt`
- `ViewListItemsEffect.kt`
- `ViewListItemsScreen.kt`
- `ViewListItemsUiState.kt`
- `ViewListItemsViewModel.kt`

New files for the list feature include:
- `EditListContract.kt`
- `ListEffect.kt`
- `ListUiState.kt`
- `ListViewModel.kt`