hotfix: remove key from MediaLazyVerticalGrid items
This commit removes the explicit key definition from the `items` block within the `MediaLazyVerticalGrid` composable. The default indexing behavior is now used for item keys.

This is being removed as it was causing a duplicate key crash in the lazy grid when a media appears more than once, hence having the same id.