refactor: update snackbar exit animation
This commit updates the exit animation for the snackbar in the `designSystem` module.
The animation now uses a spring animation with medium-low stiffness instead of a tween animation.

Specifically, the `exitAnimation` in `SnackBarAnimationConfig.kt` was changed from:
`slideOutVertically(targetOffsetY = { -2 * it }, animationSpec = tween(durationMillis))`
to:
`slideOutVertically(targetOffsetY = { -2 * it }, animationSpec = spring(stiffness = Spring.StiffnessMediumLow))`